@@ -1,109 +1,89 @@
 {
   "name": "eis-registrant",
-  "version": "1.0.0",
   "private": true,
+  "version": "1.1.0",
+  "type": "module",
   "scripts": {
-    "start-server": "cross-env NODE_ENV=development nodemon --exec babel-node server/index.js --watch server/*",
-    "serve": "cross-env NODE_ENV=production nodemon --exec babel-node server/index.js --watch server/*",
-    "serve:test": "cross-env NODE_ENV=test nodemon --exec babel-node server/index.js",
-    "start-front": "react-scripts start",
+    "dev": "vite",
+    "dev:server": "cross-env NODE_ENV=development nodemon server/index.js --watch server",
+    "start": "concurrently \"npm run dev:server\" \"npm run dev\" --kill-others --kill-others-on-fail",
     "build": "node ./buildScript",
+    "serve": "cross-env NODE_ENV=production node server/index.js",
     "lint": "npx eslint --fix --ext jsx,js src server",
-    "test": "npm run lint && react-scripts test",
+    "serve:test": "cross-env NODE_ENV=test nodemon server/index.js",
+    "test": "npm run lint && vitest run",
     "test:e2e": "concurrently \"npm run serve:test\" \"cypress run\" --kill-others --success first",
-    "start": "concurrently \"npm run start-server\" \"npm run start-front\" --kill-others --kill-others-on-fail",
     "deploy": "npm run test && npm run build && npm run test:e2e && npm run serve"
   },
   "dependencies": {
-    "@babel/polyfill": "^7.10.1",
-    "@babel/runtime": "^7.10.1",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
-    "axios": "^0.21.2",
-    "body-parser": "^1.19.0",
-    "capitalize": "^2.0.3",
-    "compression": "^1.7.4",
-    "cookie-parser": "^1.4.5",
-    "cookie-session": "^1.4.0",
-    "dompurify": "^2.2.9",
-    "dotenv": "^8.2.0",
-    "express": "^4.17.1",
-    "express-winston": "^4.0.5",
-    "grant-express": "^4.7.0",
-    "helmet": "^3.22.0",
-    "jsonwebtoken": "^9.0.0",
-    "jwk-to-pem": "^2.0.3",
-    "moment": "^2.26.0",
-    "qs": "^6.9.4",
-    "react-ga": "^3.3.0",
-    "vanilla-cookieconsent": "^2.9.2",
-    "winston": "^3.2.1",
-    "winston-daily-rotate-file": "^4.5.0"
+    "@reduxjs/toolkit": "^2.5.0",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.1.0",
+    "@testing-library/user-event": "^14.5.2",
+    "axios": "^1.7.9",
+    "body-parser": "^1.20.3",
+    "capitalize": "^2.0.4",
+    "compression": "^1.7.5",
+    "cookie-parser": "^1.4.7",
+    "cookie-session": "^2.1.0",
+    "core-js": "^3.39.0",
+    "dompurify": "^3.2.3",
+    "dotenv": "^16.4.7",
+    "express": "^4.21.2",
+    "grant-express": "^5.2.0",
+    "helmet": "^8.0.0",
+    "jsonwebtoken": "^9.0.2",
+    "jwk-to-pem": "^2.0.7",
+    "prop-types": "^15.8.1",
+    "react": "18.2.0",
+    "react-cookie": "^7.2.2",
+    "react-dom": "18.2.0",
+    "react-ga": "^3.3.1",
+    "react-helmet-async": "^2.0.5",
+    "react-layout-masonry": "^1.2.0",
+    "react-router": "^7.0.2",
+    "regenerator-runtime": "^0.14.1",
+    "semantic-ui-react": "^3.0.0-beta.2",
+    "universal-cookie": "^7.2.2",
+    "vanilla-cookieconsent": "^3.0.1",
+    "winston": "^3.17.0",
+    "winston-daily-rotate-file": "^5.0.0"
   },
   "devDependencies": {
-    "@babel/cli": "^7.11.6",
-    "@babel/core": "^7.11.6",
-    "@babel/node": "^7.10.5",
-    "@babel/preset-env": "^7.11.5",
-    "autoprefixer": "^10.2.5",
-    "canvg": "^3.0.7",
-    "cheerio": "^1.0.0-rc.3",
-    "classnames": "^2.2.6",
-    "concurrently": "^5.3.0",
-    "copyfiles": "^2.3.0",
-    "cross-env": "^5.2.1",
-    "cypress": "^7.4.0",
+    "@eslint/js": "^9.15.0",
+    "@types/react": "^19.0.1",
+    "@types/react-dom": "^19.0.1",
+    "@vitejs/plugin-react": "^4.3.4",
+    "@vitest/coverage-v8": "^2.1.8",
+    "classnames": "^2.5.1",
+    "concurrently": "^9.1.0",
+    "cross-env": "^7.0.3",
+    "cypress": "^13.16.1",
     "eis-registrant-theme": "git+https://github.com/internetee/registrant_center_theme.git",
-    "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.4",
-    "eslint": "^7.11.0",
-    "eslint-config-airbnb": "^18.2.0",
-    "eslint-config-prettier": "^6.11.0",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-jsx-a11y": "^6.3.1",
-    "eslint-plugin-prettier": "^3.1.4",
-    "eslint-plugin-react": "^7.21.4",
-    "eslint-plugin-react-hooks": "^4.0.0",
-    "flatpickr": "^4.6.3",
-    "lint-staged": "^10.3.0",
-    "nodemon": "^2.0.20",
-    "pdfmake": "^0.1.65",
-    "popper.js": "^1.16.1",
-    "prettier": "^2.1.1",
-    "prop-types": "^15.7.2",
-    "react": "^16.13.1",
-    "react-cookie": "3.1.2",
-    "react-csv": "^2.0.3",
-    "react-dom": "^16.13.1",
-    "react-flatpickr": "^3.10.0",
-    "react-helmet": "^5.2.1",
-    "react-intl": "^5.3.2",
-    "react-masonry-component": "^6.2.1",
-    "react-redux": "^7.2.0",
-    "react-responsive": "^6.1.2",
-    "react-router-dom": "5.2.0",
-    "react-scripts": "^4.0.3",
-    "redux": "^4.0.5",
-    "redux-devtools-extension": "^2.13.8",
-    "redux-mock-store": "^1.5.4",
-    "redux-state-sync": "^3.1.1",
-    "redux-thunk": "2.3.0",
-    "sass": "^1.55.0",
-    "semantic-ui-react": "0.82.5",
-    "serve-favicon": "^2.5.0"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "proxy": "https://localhost:1234"
+    "eslint": "^8.57.1",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-import": "^2.31.0",
+    "eslint-plugin-jsx-a11y": "^6.10.2",
+    "eslint-plugin-prettier": "^5.2.1",
+    "eslint-plugin-react": "^7.37.2",
+    "eslint-plugin-react-hooks": "^4.6.2",
+    "eslint-plugin-react-refresh": "^0.4.16",
+    "express-winston": "^4.2.0",
+    "fs-extra": "^11.2.0",
+    "globals": "^15.12.0",
+    "nodemon": "^3.1.7",
+    "pdfmake": "^0.2.15",
+    "react-csv": "^2.2.2",
+    "react-flatpickr": "^3.10.13",
+    "react-intl": "^7.0.4",
+    "react-redux": "^9.2.0",
+    "react-responsive": "^10.0.0",
+    "redux": "^5.0.1",
+    "redux-state-sync": "^3.1.4",
+    "sass": "^1.82.0",
+    "serve-favicon": "^2.5.0",
+    "vite": "^6.0.1",
+    "vitest": "^2.1.8"
+  }
 }
