@@ -1,44 +1,49 @@
 /* eslint-disable sort-keys */
-import axios from 'axios';
 import dotenv from 'dotenv';
+dotenv.config();
+
+import axios from 'axios';
+import banner from './utils/banner.js';
 import express from 'express';
 import helmet from 'helmet';
-import session from 'cookie-session';
-import bodyParser from 'body-parser';
-import grant from 'grant-express';
-import cookieParser from 'cookie-parser';
-import https from 'https';
-import favicon from 'serve-favicon';
-import compression from 'compression';
+import { fileURLToPath } from 'url';
+import { dirname } from 'path';
 import path from 'path';
+import grant from 'grant-express';
 import fs from 'fs';
+import bodyParser from 'body-parser';
+import cookieParser from 'cookie-parser';
 import expressWinston from 'express-winston';
 import jwkToPem from 'jwk-to-pem';
-import callbackPage from './routes/callbackPageRoute';
-import banner from './utils/banner';
-import API from './routes/apiRoute';
-import { accessLog, consoleLog, errorLog } from './utils/logger';
+import callbackPage from './routes/callbackPageRoute.js';
+import { accessLog, consoleLog, errorLog } from './utils/logger.js';
+import compression from 'compression';
+import favicon from 'serve-favicon';
+import session from 'cookie-session';
+import API from './routes/apiRoute.js';
+import https from 'https';
 
-dotenv.config();
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
 
 const privateKey = fs.readFileSync('./server.key', 'utf8');
 const certificate = fs.readFileSync('./server.crt', 'utf8');
 const credentials = { key: privateKey, cert: certificate };
 
 const {
-    AUTH_PATH,
-    CLIENT_ID,
-    CLIENT_SECRET,
-    HOST,
-    ISSUER_URL,
-    JWKS_PATH,
-    NODE_ENV,
-    REACT_APP_SERVER_PORT,
-    REDIRECT_URL,
-    SESSION_SECRET,
-    TOKEN_PATH,
-    REACT_APP_SCOPE,
-    RESPONSE_TYPE,
+  AUTH_PATH,
+  CLIENT_ID,
+  CLIENT_SECRET,
+  HOST,
+  ISSUER_URL,
+  JWKS_PATH,
+  NODE_ENV,
+  VITE_SERVER_PORT,
+  REDIRECT_URL,
+  SESSION_SECRET,
+  TOKEN_PATH,
+  VITE_SCOPE,
+  RESPONSE_TYPE,
 } = process.env;
 
 let publicKey = '';
@@ -62,27 +67,35 @@ app.use(compression()); // GZip compress responses
 
 // static files
 if (NODE_ENV !== 'development') {
-    app.use(express.static(path.join(__dirname, 'build')));
+  app.use(express.static(path.join(__dirname, 'dist')));
 }
 app.use(favicon(path.join(__dirname, '../public/favicon.ico')));
 
 app.use(
-    session({
-        httpOnly: true,
-        maxAge: 7200000,
-        secret: SESSION_SECRET,
-        secure: true,
-    })
+  session({
+      name: 'session',
+      keys: [SESSION_SECRET],
+      httpOnly: true,
+      maxAge: 7200000,
+      secure: true,
+  })
 );
 
 (async () => {
-    try {
-        const { data } = await axios.get(ISSUER_URL + JWKS_PATH);
-        console.log('Received public key from TARA'); // eslint-disable-line no-console
-        publicKey = data.keys[0]; // eslint-disable-line prefer-destructuring
-    } catch (e) {
-        console.log(`Public key request error: ${e}`); // eslint-disable-line no-console
-    }
+  try {
+      const { data } = await axios.get(ISSUER_URL + JWKS_PATH);
+      if (!data?.keys?.[0]) {
+          throw new Error('Invalid JWKS response format');
+      }
+      publicKey = data.keys[0];
+      console.log(
+        'Successfully initialized public key from eeID! -> ' + ISSUER_URL + JWKS_PATH
+      );
+  } catch (e) {
+      console.error(`Failed to fetch public key: ${e}`);
+      // Optionally, you might want to retry the fetch after a delay
+      // setTimeout(() => { /* retry logic */ }, 5000);
+  }
 })();
 
 // middlewares
@@ -94,44 +107,42 @@ app.use((req, res, next) => {
 
 const redirect_uri =
     NODE_ENV === 'development'
-        ? `https://${HOST}:${REACT_APP_SERVER_PORT}${REDIRECT_URL}`
+        ? `https://${HOST}:${VITE_SERVER_PORT}${REDIRECT_URL}`
         : `https://${HOST}${REDIRECT_URL}`;
 
 // grant auth
 const grantConfig = {
-    defaults: {
-        protocol: 'https',
-        host: HOST,
-        state: true,
-        callback: '/auth/callback',
-        transport: 'querystring',
-    },
-    openid: {
-        authorize_url: ISSUER_URL + AUTH_PATH,
-        access_url: ISSUER_URL + TOKEN_PATH,
-        oauth: 2,
-        key: CLIENT_ID,
-        secret: CLIENT_SECRET,
-        scope: REACT_APP_SCOPE,
-        redirect_uri,
-        response_type: RESPONSE_TYPE,
-        callback: REDIRECT_URL,
-        custom_params: {
-            ui_locales: LOCALE,
+  defaults: {
+      protocol: 'https',
+      host: HOST,
+      state: true,
+      callback: '/auth/callback',
+      transport: 'querystring',
+  },
+  openid: {
+      authorize_url: ISSUER_URL + AUTH_PATH,
+      access_url: ISSUER_URL + TOKEN_PATH,
+      oauth: 2,
+      key: CLIENT_ID,
+      secret: CLIENT_SECRET,
+      scope: VITE_SCOPE,
+      redirect_uri,
+      response_type: RESPONSE_TYPE,
+      callback: REDIRECT_URL,
+      overrides: {
+        en: {
+          custom_params: { ui_locales: 'en' }
         },
-    }
+        et: {
+          custom_params: { ui_locales: 'et' }
+        }
+      }
+  }
 };
 
-if (REACT_APP_SCOPE.includes('webauthn')) {
-    grantConfig.openid.scope = REACT_APP_SCOPE.replace(/(?:^|\s)webauthn(?:\s|$)/, ' ').trim();
-    grantConfig.openid.overrides = {
-        webauthn: {
-            scope: REACT_APP_SCOPE,
-        },
-    };
-  }
-  
-app.use(grant(grantConfig));
+// Create grant middleware instance
+const grantMiddleware = grant(grantConfig);
+app.use(grantMiddleware);
 
 app.use(helmet());
 // api
@@ -140,7 +151,10 @@ app.get('/api/user', API.getUser);
 app.post('/api/destroy', API.destroyUser);
 app.get('/api/confirms/:name/:type/:token', API.getRegistrantUpdate);
 app.post('/api/confirms/:name/:type/:token/:action', API.sendVerificationStatus);
+
 app.all('/api/*', API.checkAuth);
+
+// Protected routes below
 app.get('/api/domains', API.getDomains);
 app.get('/api/domains/:uuid', API.getDomain);
 app.post('/api/domains/:uuid/registry_lock', API.setDomainRegistryLock);
@@ -151,15 +165,19 @@ app.get('/api/companies', API.getCompanies);
 app.get('/api/contacts', API.getContacts);
 app.get('/api/contacts/:uuid', API.getContacts);
 app.patch('/api/contacts/:uuid', API.setContact);
+// Basic health check endpoint
+app.get('/api/health', (req, res) => {
+  res.json({ status: 'ok' });
+});
 
 // all page rendering
 app.get(REDIRECT_URL, (req, res) => callbackPage(req, res, jwkToPem(publicKey).trim()));
 
-app.get('/*', (req, res) => res.sendFile(path.join(__dirname, 'build', 'index.html')));
+app.get('/*', (req, res) => res.sendFile(path.join(__dirname, 'dist', 'index.html')));
 
 const server = https
     .createServer(credentials, app)
-    .listen(NODE_ENV === 'test' ? 4000 : REACT_APP_SERVER_PORT, () => {
+    .listen(NODE_ENV === 'test' ? 4000 : VITE_SERVER_PORT, () => {
         banner();
         // eslint-disable-next-line no-console
         console.log(`Environment: ${NODE_ENV}`);
