@@ -1,9 +1,8 @@
-import React from 'react';
 import { Icon } from 'semantic-ui-react';
 import { FormattedMessage } from 'react-intl';
 import { connect } from 'react-redux';
 import dompurify from 'dompurify';
-
+import * as CookieConsent from "vanilla-cookieconsent";
 import PropTypes from 'prop-types';
 
 const sanitizer = dompurify.sanitize;
@@ -111,9 +110,7 @@ const mapStateToProps = ({ ui }) => ({
 });
 
 const showCookieSettings = () => {
-    if (window.CC) {
-        window.CC.showSettings(1);
-    }
+    CookieConsent.showPreferences();
 };
 
 export default connect(mapStateToProps)(MainFooter);
