@@ -9,7 +9,5 @@ export { default as PageMessage } from './PageMessage/PageMessage';
 export { default as Roles } from './Roles/Roles';
 export { default as WhoIsConfirmDialog } from './WhoIsConfirmDialog/WhoIsConfirmDialog';
 export { default as ProtectedRoute } from './ProtectedRoute/ProtectedRoute';
-export { default as CookieConsent } from './common/CookieConsent/CookieConsent';
+// export { default as CookieConsent } from './common/CookieConsent/CookieConsent';
 // pass through common 3rd party components to also cut down on import boilerplate
-
-export { Helmet } from 'react-helmet';
