@@ -1,98 +1,59 @@
 /* eslint-disable */
 import axios from 'axios';
-import https from 'https';
-
-const agent = new https.Agent({
-    rejectUnauthorized: false,
-});
 
 const instance = axios.create({
     headers: {
         Accept: 'application/json',
         'Content-Type': 'application/json',
     },
-    httpsAgent: agent,
+    withCredentials: true,
     validateStatus: (status) => status >= 200 && status < 300,
 });
 
+// Response interceptor for error handling
+// instance.interceptors.response.use(
+//     response => response,
+//     error => {
+//         console.error('API Error:', error);
+//         return Promise.reject(error);
+//     }
+// );
+
 export default {
     fetchMenu: (type) => {
-        return axios({
-            credentials: 'include',
-            headers: {
-                Accept: 'application/json',
-                'Content-Type': 'application/json',
-            },
-            httpsAgent: agent,
-            method: 'GET',
-            url: `/api/menu/${type}`,
-        });
+        return instance.get(`/api/menu/${type}`);
     },
+
     fetchUser: () => {
-        return axios.get('/api/user', {
-            credentials: 'include',
-            headers: {
-                Accept: 'application/json',
-                'Content-Type': 'application/json',
-            },
-            httpsAgent: agent,
-            method: 'GET',
+        return instance.get('/api/user', {
             timeout: 15000,
         });
     },
 
     destroyUser: () => {
-        return axios.post(
-            '/api/destroy',
-            {},
-            {
-                credentials: 'include',
-                headers: {
-                    Accept: 'application/json',
-                    'Content-Type': 'application/json',
-                },
-                httpsAgent: agent,
-            }
-        );
+        return instance.post('/api/destroy');
     },
 
     fetchVerification: (domain, token, type) => {
         return instance.get(`/api/confirms/${domain}/${type}/${token}`, {
-            credentials: 'include',
-            method: 'GET',
             timeout: 4000,
         });
     },
 
     sendVerificationResponse: (name, token, action, type) => {
         return instance.post(`/api/confirms/${name}/${type}/${token}/${action}`, {
-            credentials: 'include',
-            method: 'POST',
             timeout: 4000,
         });
     },
 
     fetchUpdateContacts: (uuid) => {
         return instance.get(`/api/contacts/${uuid}/do_need_update_contacts`, {
-            credentials: 'include',
-            method: 'GET',
             timeout: 4000,
         });
     },
 
     updateContacts: (uuid) => {
-        return axios.post(
-            `/api/contacts/${uuid}/update_contacts`,
-            {},
-            {
-                credentials: 'include',
-                headers: {
-                    Accept: 'application/json',
-                    'Content-Type': 'application/json',
-                },
-                httpsAgent: agent,
-            }
-        );
+        return instance.post(`/api/contacts/${uuid}/update_contacts`);
     },
 
     fetchDomains: (uuid = false, offset = 0, simplify = true, tech = 'init') => {
@@ -122,7 +83,13 @@ export default {
         }
         return instance.get(`/api/contacts?offset=${offset}`);
     },
-    updateContact: (uuid, form) => instance.patch(`/api/contacts/${uuid}`, JSON.stringify(form)),
-    setDomainRegistryLock: (uuid, extensionsProhibited) => instance.post(`/api/domains/${uuid}/registry_lock?extensionsProhibited=${extensionsProhibited}`),
-    deleteDomainRegistryLock: (uuid) => instance.delete(`/api/domains/${uuid}/registry_lock`),
+
+    updateContact: (uuid, form) => 
+        instance.patch(`/api/contacts/${uuid}`, JSON.stringify(form)),
+
+    setDomainRegistryLock: (uuid, extensionsProhibited) => 
+        instance.post(`/api/domains/${uuid}/registry_lock?extensionsProhibited=${extensionsProhibited}`),
+
+    deleteDomainRegistryLock: (uuid) => 
+        instance.delete(`/api/domains/${uuid}/registry_lock`),
 };
